Resolved pylint for duplication and removed the redundant conftest
